+++ apps/app/src/main.ts
@@ -1,5 +1,5 @@
-import { createApp } from 'vue';
 import { createProvider, LocalService, createModules } from '@vtj/web';
+import { createApp } from 'vue';
 import router from './router';
 import App from './App.vue';
 import './style/index.scss';