+++ dev/src/extension/index.ts
@@ -1,5 +1,7 @@
-import { widgetManager, type AppWidget } from '@vtj/pro';
+import { widgetManager, type AppWidget, type ExtensionFactory } from '@vtj/pro';
+import { type App } from 'vue';
 import { VtjIconDatabase } from '@vtj/icons';
+import { request, jsonp } from '@vtj/utils';
 import Meta from './Meta.vue';
 
 const meta: AppWidget = {
@@ -14,4 +16,21 @@ const meta: AppWidget = {
 };
 widgetManager.register(meta);
 
-export default function () {}
+request.useRequest((req: any) => {
+  req.headers = Object.assign({}, req.headers || {}, { abc: 'abc' });
+  return req;
+});
+
+const factory: ExtensionFactory = () => {
+  return {
+    adapter: {
+      request,
+      jsonp
+    },
+    install(app: App, engine: any) {
+      console.log('ExtensionFactory install', app, engine);
+    }
+  };
+};
+
+export default factory;