+++ apps/library/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.10](https://gitee.com/newgateway/vtj/compare/vtj-project-library@0.8.9...vtj-project-library@0.8.10) (2024-05-07)
+
+**Note:** Version bump only for package vtj-project-library
+
+
+
+
+
+## [0.8.9](https://gitee.com/newgateway/vtj/compare/vtj-project-library@0.8.8...vtj-project-library@0.8.9) (2024-05-07)
+
+**Note:** Version bump only for package vtj-project-library
+
+
+
+
+
+## [0.8.8](https://gitee.com/newgateway/vtj/compare/vtj-project-library@0.8.7...vtj-project-library@0.8.8) (2024-05-06)
+
+**Note:** Version bump only for package vtj-project-library
+
+
+
+
+
+## [0.8.7](https://gitee.com/newgateway/vtj/compare/vtj-project-library@0.8.6...vtj-project-library@0.8.7) (2024-05-06)
+
+**Note:** Version bump only for package vtj-project-library
+
+
+
+
+
+## [0.8.6](https://gitee.com/newgateway/vtj/compare/vtj-project-library@0.8.5...vtj-project-library@0.8.6) (2024-05-06)
+
+**Note:** Version bump only for package vtj-project-library
+
+
+
+
+
+## [0.8.5](https://gitee.com/newgateway/vtj/compare/vtj-project-library@0.8.4...vtj-project-library@0.8.5) (2024-05-04)
+
+**Note:** Version bump only for package vtj-project-library
+
+
+
+
+
+## [0.8.4](https://gitee.com/newgateway/vtj/compare/vtj-project-library@0.8.3...vtj-project-library@0.8.4) (2024-05-01)
+
+**Note:** Version bump only for package vtj-project-library
+
+
+
+
+
+
 ## [0.8.3](https://gitee.com/newgateway/vtj/compare/vtj-project-library@0.8.2...vtj-project-library@0.8.3) (2024-04-27)
 
 **Note:** Version bump only for package vtj-project-library