+++ lcdp/CHANGELOG.md
@@ -3,6 +3,121 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.25](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.24...lcdp@0.8.25) (2024-05-07)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
+## [0.8.24](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.23...lcdp@0.8.24) (2024-05-07)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
+## [0.8.23](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.22...lcdp@0.8.23) (2024-05-07)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
+## [0.8.22](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.21...lcdp@0.8.22) (2024-05-07)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
+## [0.8.21](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.20...lcdp@0.8.21) (2024-05-06)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
+## [0.8.20](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.19...lcdp@0.8.20) (2024-05-06)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
+## [0.8.19](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.18...lcdp@0.8.19) (2024-05-06)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
+## [0.8.18](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.17...lcdp@0.8.18) (2024-05-04)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
+## [0.8.17](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.16...lcdp@0.8.17) (2024-05-02)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
+## [0.8.16](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.15...lcdp@0.8.16) (2024-05-02)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
+## [0.8.15](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.14...lcdp@0.8.15) (2024-05-02)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
+## [0.8.14](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.13...lcdp@0.8.14) (2024-05-02)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
+## [0.8.13](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.12...lcdp@0.8.13) (2024-05-02)
+
+
+### Bug Fixes
+
+* 🐛 lcdp favicon ([eb2bade](https://gitee.com/newgateway/vtj/commits/eb2badeef5d0a5ed9854b00102a60525848f702b))
+
+
+
+
+
+## [0.8.12](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.11...lcdp@0.8.12) (2024-05-01)
+
+**Note:** Version bump only for package lcdp
+
+
+
+
+
 ## [0.8.11](https://gitee.com/newgateway/vtj/compare/lcdp@0.8.10...lcdp@0.8.11) (2024-05-01)
 
 **Note:** Version bump only for package lcdp