+++ packages/charts/CHANGELOG.md
@@ -3,6 +3,121 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.25](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.24...@vtj/charts@0.8.25) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
+## [0.8.24](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.23...@vtj/charts@0.8.24) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
+## [0.8.23](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.22...@vtj/charts@0.8.23) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
+## [0.8.22](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.21...@vtj/charts@0.8.22) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
+## [0.8.21](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.20...@vtj/charts@0.8.21) (2024-05-06)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
+## [0.8.20](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.19...@vtj/charts@0.8.20) (2024-05-06)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
+## [0.8.19](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.18...@vtj/charts@0.8.19) (2024-05-06)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
+## [0.8.18](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.17...@vtj/charts@0.8.18) (2024-05-04)
+
+
+### Bug Fixes
+
+* 🐛 指令引用计算属性出码错误 ([9e7e81b](https://gitee.com/newgateway/vtj/commits/9e7e81b7a3d96a81140121954ae6b22846f4d611))
+
+
+
+
+
+## [0.8.17](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.16...@vtj/charts@0.8.17) (2024-05-02)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
+## [0.8.16](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.15...@vtj/charts@0.8.16) (2024-05-02)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
+## [0.8.15](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.14...@vtj/charts@0.8.15) (2024-05-02)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
+## [0.8.14](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.13...@vtj/charts@0.8.14) (2024-05-02)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
+## [0.8.13](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.12...@vtj/charts@0.8.13) (2024-05-02)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
+## [0.8.12](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.11...@vtj/charts@0.8.12) (2024-05-01)
+
+**Note:** Version bump only for package @vtj/charts
+
+
+
+
+
 ## [0.8.11](https://gitee.com/newgateway/vtj/compare/@vtj/charts@0.8.10...@vtj/charts@0.8.11) (2024-05-01)
 
 **Note:** Version bump only for package @vtj/charts