+++ packages/designer/CHANGELOG.md
@@ -3,6 +3,127 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.25](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.24...@vtj/designer@0.8.25) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.8.24](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.23...@vtj/designer@0.8.24) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.8.23](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.22...@vtj/designer@0.8.23) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.8.22](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.21...@vtj/designer@0.8.22) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.8.21](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.20...@vtj/designer@0.8.21) (2024-05-06)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.8.20](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.19...@vtj/designer@0.8.20) (2024-05-06)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.8.19](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.18...@vtj/designer@0.8.19) (2024-05-06)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.8.18](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.17...@vtj/designer@0.8.18) (2024-05-04)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.8.17](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.16...@vtj/designer@0.8.17) (2024-05-02)
+
+
+### Features
+
+* ✨ extension add params ([ce517b8](https://gitee.com/newgateway/vtj/commits/ce517b8075a491186fc69ae8a757d810022538d4))
+
+
+
+
+
+## [0.8.16](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.15...@vtj/designer@0.8.16) (2024-05-02)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.8.15](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.14...@vtj/designer@0.8.15) (2024-05-02)
+
+
+### Bug Fixes
+
+* 🐛 engine 处理 adapter ([03478a9](https://gitee.com/newgateway/vtj/commits/03478a9ce700dddcaccc5f5922f4e3391ad001b2))
+
+
+
+
+
+## [0.8.14](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.13...@vtj/designer@0.8.14) (2024-05-02)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
+## [0.8.13](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.12...@vtj/designer@0.8.13) (2024-05-02)
+
+
+### Features
+
+* ✨ 插件增加install参数 ([45b704b](https://gitee.com/newgateway/vtj/commits/45b704b445bf7f7c935844a2ee53b5714f4f2140))
+
+
+
+
+
+## [0.8.12](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.11...@vtj/designer@0.8.12) (2024-05-01)
+
+**Note:** Version bump only for package @vtj/designer
+
+
+
+
+
 ## [0.8.11](https://gitee.com/newgateway/vtj/compare/@vtj/designer@0.8.10...@vtj/designer@0.8.11) (2024-05-01)
 
 **Note:** Version bump only for package @vtj/designer