+++ platforms/web/CHANGELOG.md
@@ -3,6 +3,121 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.25](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.24...@vtj/web@0.8.25) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.24](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.23...@vtj/web@0.8.24) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.23](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.22...@vtj/web@0.8.23) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.22](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.21...@vtj/web@0.8.22) (2024-05-07)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.21](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.20...@vtj/web@0.8.21) (2024-05-06)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.20](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.19...@vtj/web@0.8.20) (2024-05-06)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.19](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.18...@vtj/web@0.8.19) (2024-05-06)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.18](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.17...@vtj/web@0.8.18) (2024-05-04)
+
+
+### Features
+
+* ✨ cli自动设置buildTarget ([c0e626b](https://gitee.com/newgateway/vtj/commits/c0e626bbe099c224b8cda1ca04b86a4a8b018dda))
+
+
+
+
+
+## [0.8.17](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.16...@vtj/web@0.8.17) (2024-05-02)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.16](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.15...@vtj/web@0.8.16) (2024-05-02)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.15](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.14...@vtj/web@0.8.15) (2024-05-02)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.14](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.13...@vtj/web@0.8.14) (2024-05-02)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.13](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.12...@vtj/web@0.8.13) (2024-05-02)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
+## [0.8.12](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.11...@vtj/web@0.8.12) (2024-05-01)
+
+**Note:** Version bump only for package @vtj/web
+
+
+
+
+
 ## [0.8.11](https://gitee.com/newgateway/vtj/compare/@vtj/web@0.8.10...@vtj/web@0.8.11) (2024-05-01)
 
 **Note:** Version bump only for package @vtj/web